# dist/openapi.yaml
openapi: 3.0.3
servers:
  # # Added by API Auto Mocking Plugin
  # # Added by API Auto Mocking Plugin
  # # Added by API Auto Mocking Plugin


  # - description: SwaggerHub API Auto Mocking
  #   url: https://virtserver.swaggerhub.com/bengutionuneversity/demo/1.0.0
  # - description: SwaggerHub API Auto Mocking
  #   url: https://virtserver.swaggerhub.com/bgu-ea6/APIProject/1.0.0
  # - description: SwaggerHub API Auto Mocking
  #   url: https://virtserver.swaggerhub.com/bgu-ea6/RecepiesAPI/1.0.0

  servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://virtserver.swaggerhub.com/bengutionuneversity/demo/1.0.0
    description: SwaggerHub API Auto Mocking
info:
  description: This is an API for our recepies website.
  version: "1.0.0"
  title: Recepies API
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /recipes:
    get:
      tags:
        - recipe
      summary: Get 3 random recipe previews for homepage
      operationId: getRandomHomepageRecipes
      description: |
        Returns a catalog of 3 random recipe previews to be displayed on the homepage.  
        This endpoint is used for the initial homepage display.  
        A "Load More" button on the UI can trigger this endpoint again to fetch a different set of random recipes.
        *Currentlly not support on recipes of users, only recipes from spooncular catalog - which means no U_[ID] recipes!
      responses:
        '200':
          description: A list of 3 random recipe previews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
        '500':
          description: Server error while retrieving recipes
...
  /user/recipes:
    get:
      tags: 
        - user
      summary: Get created recipes (preview) of the logged-in user
      operationId: getUserRecipes
      description: Returns all recipes (preview) created by the currently authenticated user.
      responses:
        '200':
          description: A list of the user's created recipes (preview)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
        '401':
          description: 'Unauthorized: user not found'
          
    post:
      tags:
        - user
      summary: Add a recipe to the user's resepies list
      operationId: addRecipeToList
      description: Adds a new recipe to the user's recepies list.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
        description: The recepie details for addition
      responses:
        '200':
          description: Recipe added to list
        '400':
          description: Invalid recipe details
        '401':
          description: 'Unauthorized: user not found'
  /user/recipes/{recipeId}:
    delete:
      tags:
        - user
      summary: Remove a user-created recipe. Here you need to insert only the number of recipe ans system will validate the U_[ID] part.
      operationId: deleteUserRecipe
      description: Deletes a recipe created by the user.
      parameters:
        - name: recipeId
          in: path
          required: true
          description: The unique ID of the recipe to delete.
          schema:
            type: string
            example: 1
      responses:
        '200':
          description: Recipe deleted successfully
        '401':
          description: 'Unauthorized: user not found'
        '404':
          description: Recipe not found or doesn't belong to the user
        '400':
          description: Invalid recipe ID format
  
  # /user/watch:
  #   post:
  #     tags:
  #       - user
  #     summary: Mark a recipe as watched by current session user. Please if you want to add a user's recipes to watch - add a U_[id] next to it.
  #     operationId: markWatchedRecipe
  #     description: Marks a recipe as watched by the current session user.
  #     requestBody:
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             type: object
  #             required:
  #               - recipeId
  #             properties:
  #               recipeId:
  #                 type: integer
  #                 description: ID of the recipe to mark as watched.
  #                 example: 125 or U_125
  #     responses:
  #       '200':
  #         description: Recipe marked as watched
  #       '404':
  #         description: Recipe not found
  #       '401':
  #         description: "Unauthorized: user not found"

  #   get:
  #     tags:
  #       - user
  #     summary: Get 3 watched recipes by current user
  #     operationId: getWatchedRecipe
  #     description: Get 3 watched recipes by current user
  #     responses:
  #       '200':
  #         description: A list of up to 3 watched recipes by the current user.
  #         content:
  #           application/json:
  #             schema:
  #               type: array
  #               items:
  #                 $ref: '#/components/schemas/Recipe'
  #       '404':
  #         description: No watched recipes found for the user.
  #       '401':
  #         description: 'Unauthorized: user not found'
  /user/favorites:
    post:
      tags:
        - favorites
      summary: Add an exist recipe to the user's favorites list
      operationId: addRecipeToFavorites
      description: Adds a recipe (from catalog or from user's created recipes list) to the user's favorites list.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - recipe_id
              properties:
                recipe_id:
                  type: integer
                  description: Internal identifier of the recipe to add.
                  example: "638893"
      responses:
        '200':
          description: Recipe added to favorites
        '400':
          description: Invalid recipe ID
        '404':
          description: Recipe not found
        '401':
          description: 'Unauthorized: user not found'
    get:
      tags:
        - favorites
      summary: Get the list of user's favorite recipes
      operationId: getUserFavorites
      description: Retrieves all the recipes that the user has added to their favorites (prevew).
      responses:
        '200':
          description: A list of the user's favorite recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
        '401':
          description: 'Unauthorized: user not found'
        '404':
          description: No favorites found for the user
  /user/favorites/{recipeId}:
    delete:
      tags:
        - favorites
      summary: Remove a favorited recipe from favorites
      operationId: deleteFavRecipe
      description: Remove a favorited recipe from favorites
      parameters:
        - name: recipeId
          in: path
          required: true
          description: The unique ID of the recipe to delete.
          schema:
            type: string
            example: U_123
      responses:
        '200':
          description: Recipe deleted successfully
        '401':
          description: 'Unauthorized: user not found'
        '404':
          description: Recipe not found or doesn't favorited by the user
        '400':
          description: Invalid recipe ID format
  /user/recipes/family:
    get:
      tags: 
        - user
      summary: Get recepies by the developers families
      operationId: getFamilyRecipes
      description: Returns all recipes created by the developers of this website.
      responses:
        '200':
          description:  A list of the developers recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FamilyRecipe'
        '401':
          description: Unauthorized – user must be logged in to access this resource
        '404':
          description: No family recipes found for this user

# Schemas:      
components:
  schemas:
    Recipe:
      type: object
      required:
        - title
        - image
        - readyInMinutes
        - popularity
        - vegan,
        - vegetarian,
        - glutenFree
      properties:
        # recipeID:
        #   type: string
        #   description: Unique identifier (internal in the system's DB with a 'U' prefix) for the recipe, 'U_{Incremental number}'.
        #   example: "U_125 or 125"
        title:
          type: string
          description: Full name of the recipe.
          minLength: 2
          maxLength: 100
          example: Pasta Rose
        image:
          type: string
          format: uri
          description: URL to an image representing the recipe.
          example: https://s.lightorangebean.com/media/20240914160809/Spicy-Penne-Pasta_-done.png
        readyInMinutes:
          type: integer
          description: Preparation time in minutes.
          example: 15
        popularity:
          type: number
          format: integer
          description: Popularity among users will be the number of likes for this recipe.
          example: 105
        vegan:
          type: boolean
          description: True if suitable for vegans.
          example: true
        vegetarian:
          type: boolean
          description: True if suitable for vegetarians.
          example: true
        glutenFree:
          type: boolean
          description: True if gluten-free or uses alternatives.
          example: true
        extendedIngredients:
          type: array
          description: "List of ingredients in the format ingredient: amount."
          items:
            type: string
            example: "Flour: 2 cups"
        instructions:
          type: string
          description: Full preparation instructions.
          example: "Mix the ingredients and bake at 180°C for 30 minutes."
        servings:
          type: integer
          description: Number of servings the recipe makes.
          example: 4
...
    FamilyRecipe:
      type: object
      required:
        - recipeID
        - userID
        - title
        - readyInMinutes
        - image
        - extendedIngredients
        - instructions
        - timeInYear
        - recipeMaster
      properties:
        recipeID:
          type: string
          example: khachapuri001
        userID:
          type: integer
          example: 42
        title:
          type: string
          example: Traditional Khachapuri
        readyInMinutes:
          type: integer
          example: 90
        image:
          type: string
          format: uri
          example: https://example.com/images/khachapuri.jpg
        extendedIngredients:
          type: string
          example: flour, water, yeast, salt, eggs, sulguni cheese, butter
        instructions:
          type: string
          example: |
            1. Knead the dough and let it rise for about an hour.
            2. Prepare the filling by mixing the cheese and eggs.
            3. Roll out the dough, fill it, and shape it into a boat.
            4. Bake in a preheated oven at 200°C (392°F) for 25 minutes.
            5. Add an egg yolk to the center and bake for 3 more minutes.
            6. Serve hot with a pat of butter on top.
        timeInYear:
          type: string
          example: Winter
        recipeMaster:
          type: string
          example: Natalie's grandmother
...